main: Save console daemon stderr and switch to file.
It appears the nobody user cannot open syslog. We could copy the
child's stderr to our syslog writer - but we do not want to trust
any stderr from it post-setup. The next-best compromise is to
share a log file fd and have the child write to a log file.

This commit also ensures that no environment variables are shared
with the console daemon process.