main: Fixed occasional console mode hang.
ListenerCtx would randomly accept one connection, stealing it
away from the console daemon child process. This would leave
the console client process hanging because nothing would ever
read from it or write to it.

ListenerCtx was the wrong way to solve this. Instead, we now
use the new passfd.ShareableUnixSocket helper function. This
function does not attempt to accept new connections in the
background.